requirements.txt
Flask==2.0.3
gunicorn==19.9.0
pandas==1.0.3
pandas-gbq==0.17.4
google-cloud-bigquery==3.4.0
pyarrow==7.0.0
telebot==0.0.4
pyTelegramBotAPI==3.8.3
ga4mp==1.1.1
google-cloud-storage==2.3.0
google-cloud-bigquery-storage==2.4.0
itsdangerous==2.1.2
Jinja2==3.1.1
python-dateutil==2.8.2
numpy==1.16.5
joblib==1.1.0
lightgbm==3.3.2
protobuf==3.20.1
Werkzeug==2.2.2
python-dotenv==1.0.0


